requirements.txt
asgiref==3.5.2
Django==4.1.1
psycopg2-binary==2.9.3
sqlparse==0.4.2
Pillow==9.2.0
django-mptt==0.13.4
djangorestframework==3.14.0
Markdown==3.4.1
django-filter==22.1